backend_ppe/requirements.txt
fastapi
uvicorn
pillow
python-multipart
torch
opencv-python-headless
pyyaml
numpy
ultralytics  # YOLOv8 library